🚚 Rename: `CategoryButton` 파일 domains -> blocks로 이동
